Unify preprocessing functions (#1027)
* move preprocessing functions to common preprocessing util file

* update docs, rename load_preprocess to load_and_preprocess

* update preprocess init, update preproc_or_load_group in demod_mm

* move shared preprocess functions from site_pipeline/util.py to preprocess_util.py

* fix logger in update_preprocess_plots

* update demod_mapmaker logger

* incorporate update for test fix

* small cleanup

* update docstring

---------
