+++ sotodlib/preprocess/__init__.py
@@ -1,2 +1,3 @@
 from .pcore import _Preprocess, Pipeline
 from .processes import *
+from .preprocess_util import *
