+++ sotodlib/site_pipeline/preprocess_obs.py
@@ -8,38 +8,10 @@ from typing import Optional, List
 
 from sotodlib import core
 import sotodlib.site_pipeline.util as sp_util
+from sotodlib.preprocess import preprocess_util as pp_util
 from sotodlib.preprocess import _Preprocess, Pipeline, processes
 
 logger = sp_util.init_logger("preprocess")
-
-def _get_preprocess_context(configs, context=None):
-    if type(configs) == str:
-        configs = yaml.safe_load(open(configs, "r"))
-    
-    if context is None:
-        context = core.Context(configs["context_file"])
-        
-    if type(context) == str:
-        context = core.Context(context)
-    
-    # if context doesn't have the preprocess archive it in add it
-    # allows us to use same context before and after calculations
-    found=False
-    if context.get("metadata") is None:
-        context["metadata"] = []
-
-    for key in context.get("metadata"):
-        if key.get("name") == "preprocess":
-            found=True
-            break
-    if not found:
-        context["metadata"].append( 
-            {
-                "db" : configs["archive"]["index"],
-                "name" : "preprocess"
-            }
-        )
-    return configs, context
 
 def preprocess_obs(
     obs_id, 
@@ -95,7 +67,7 @@ def preprocess_obs(
     if success != 'end':
         return
 
-    policy = sp_util.ArchivePolicy.from_params(configs['archive']['policy'])
+    policy = pp_util.ArchivePolicy.from_params(configs['archive']['policy'])
     dest_file, dest_dataset = policy.get_dest(obs_id)
     logger.info(f"Saving data to {dest_file}:{dest_dataset}")
     proc_aman.save(dest_file, dest_dataset, overwrite=overwrite)
@@ -121,30 +93,6 @@ def preprocess_obs(
             lmsi.core([Path(x.name) for x in Path(new_plots).glob("*.png")],
                       Path(configs["lmsi_config"]),
                       Path(os.path.join(new_plots, 'index.html')))
-
-def load_preprocess_obs(obs_id, configs="preprocess_obs_configs.yaml", context=None ):
-    """ Loads the saved information from the preprocessing pipeline and runs the
-    processing section of the pipeline. 
-
-    Assumes preprocess_tod has already been run on the requested observation. 
-    
-    Arguments
-    ----------
-    obs_id: multiple
-        passed to `context.get_obs` to load AxisManager, see Notes for 
-        `context.get_obs`
-    configs: string or dictionary
-        config file or loaded config directory
-    """
-
-    configs, context = _get_preprocess_context(configs, context)
-    meta = load_preprocess_det_select(obs_id, configs=configs, context=context)
-    
-    pipe = Pipeline(configs["process_pipe"], logger=logger)
-    aman = context.get_obs(meta, no_signal=True)
-    pipe.run(aman, aman.preprocess)
-    return aman
-
 
 def get_parser(parser=None):
     if parser is None:
@@ -209,39 +157,13 @@ def main(
     planet_obs: bool = False,
     verbosity: Optional[int] = None,
  ):
-    configs, context = _get_preprocess_context(configs)
+    configs, context = pp_util.get_preprocess_context(configs)
     logger = sp_util.init_logger("preprocess", verbosity=verbosity)
-    if (min_ctime is None) and (update_delay is not None):
-        # If min_ctime is provided it will use that..
-        # Otherwise it will use update_delay to set min_ctime.
-        min_ctime = int(time.time()) - update_delay*86400
-
-    if obs_id is not None:
-        tot_query = f"obs_id=='{obs_id}'"
-    else:
-        tot_query = "and "
-        if min_ctime is not None:
-            tot_query += f"timestamp>={min_ctime} and "
-        if max_ctime is not None:
-            tot_query += f"timestamp<={max_ctime} and "
-        if query is not None:
-            tot_query += query + " and "
-        tot_query = tot_query[4:-4]
-        if tot_query=="":
-            tot_query="1"
     
-    if not(tags is None):
-        for i, tag in enumerate(tags):
-            tags[i] = tags[i].lower()
-            if '=' not in tags[i]:
-                tags[i] += '=1'
-
-    if planet_obs:
-        obs_list = []
-        for tag in tags:
-            obs_list.extend(context.obsdb.query(tot_query, tags=[tag]))
-    else:
-        obs_list = context.obsdb.query(tot_query, tags=tags)
+    obs_list = sp_util.get_obslist(context, query=query, obs_id=obs_id, min_ctime=min_ctime, 
+                                   max_ctime=max_ctime, update_delay=update_delay, tags=tags, 
+                                   planet_obs=planet_obs)
+    
     if len(obs_list)==0:
         logger.warning(f"No observations returned from query: {query}")
     run_list = []
